Refactor MSAL code out of AzureAD Component (#45)
Our code is very tightly coupled between our React component and the interaction with the Msal libraries. This not only makes it impossible to support other authentication frameworks in the future (or as previously discussed, making it also work with ADAL), but also makes diagnosing issues more difficult.

The general design is that there is an IAuthProvider that has 3 different methods that I expect every authentication provider will have: login, logout, and getUserInfo. There is then a subclass which is the MsalAuthProvider; an abstract class that contains all the shared logic between our two authentication methods, popup and redirect. Any method specific code is then split up into the MsalPopupAuthProvider and MsalRedirectAuthProvider.

Commits:
* Move common interfaces into its own file

* Break out MSAL business logic into seperate classes

The MsalAuthProvider is abstract base class that has some shared functionality
for all Msal type auth providers.
MsalPopupAuthProvider and MsalRedirectAuthProvider contain special
implementation details of their respectice
login methods.
AzureAD is not just the React component code instead of a mix of
authentication logic and React component code

* Remove userInfoCallback from config and set state in component

* Instantiate different auth provider based on type

* Add callback to notify consumer when user info has been updated

* Remove init and replace with constructor

* Move LoginType back to Component and fix binding error

* Move StorageLocations to MsalAuthProvider as it is only used there

* Move not-common interfaces out of Interfaces.ts

* Update tests to pass with new methods

These tests are barely verifying anything, but its better than nothing,
so for now they are back to working

* Slight optimization: make default implementation of tokenRedirectCallback no-op

* Bump minor version due to breaking changes